Asynchronous Changes.
Attempted to fix the freezing of the window, whenever a task is being proccessed. Minor changes to the GUI, being more dynamic responsive. Updated IRepository with the correct naming convention regarding Tasks. Added a View Model base to implement the INotifyPropertyChanged, so any property is updated dynamically. Also fixed an issue where the call had to manually called for whenever the data collection was modified. Should be updated automatically. Removed unused using statements inside the view models. Been reworking styles, so be located underneath each other, so all styles are located in the Windows.Resources directory instead of all over the file.